WEBAPP-314: Move .styles to component file
