+++ src/modules/common/components/Breadcrumbs.js
@@ -1,8 +1,47 @@
 // @flow
 
 import * as React from 'react'
-import { Breadcrumb, OrderedList, Wrapper } from './Breadcrumbs.styles'
 import type { UiDirectionType } from '../../app/StateType'
+import styled from 'styled-components'
+
+const opposite = direction => direction === 'ltr' ? 'rtl' : 'ltr'
+const Wrapper = styled.div`
+  margin: 10px 0;
+  text-align: end;
+  white-space: nowrap;
+  overflow-x: auto;
+  direction: ${props => opposite(props.direction)};
+`
+
+const OrderedList = styled.ol`
+  direction: ${props => props.direction};
+  display: inline-block;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+
+  /* avoid changing height when switching between pages (show one line even if there are no breadcrumbs) */
+  &:empty::after {
+    padding-left: 1px;
+    content: '';
+  }
+`
+
+const Breadcrumb = styled.li`
+  display: inline;
+  
+  & * {
+    ${props => props.theme.helpers.removeLinkHighlighting}
+    color: ${props => props.theme.colors.textSecondaryColor};
+    font-size: 15px;
+  }
+  
+  &::before {
+    color: ${props => props.theme.colors.textDecorationColor};
+    font-size: 16px;
+    content: ' > ';
+  }
+`
 
 type PropsType = {
   children: Array<React.Node>,
