--- src/modules/common/components/Failure.styles.js
@@ -1,11 +1,0 @@
-// @flow
-
-import styled from 'styled-components'
-
-export const Centered = styled.div`
-  & > * {
-    display: block;
-    margin-top: 50px;
-    text-align: center;
-  }
-`
