+++ src/modules/common/components/Headroom.js
@@ -2,7 +2,7 @@
 
 import type { Element, Node } from 'react'
 import * as React from 'react'
-import { HeaderWrapper } from './Headroom.styles'
+import styled from 'styled-components'
 
 const UPWARDS = 'up'
 const DOWNWARDS = 'down'
@@ -35,6 +35,15 @@ type StateType = {
   mode: ModeType,
   transition: boolean
 }
+
+const HeaderWrapper = styled.div`
+  position: ${props => props.positionStickyDisabled ? 'static' : 'sticky'};
+  top: ${props => props.top}px;
+  z-index: 1;
+  transform: translateY(${props => props.translateY}px);
+  ${props => props.transition && `transition: transform 0.2s ease-out;`}
+  ${props => props.static && `transition: none;`}
+`
 
 class Headroom extends React.PureComponent<PropsType, StateType> {
   static defaultProps = {
