--- src/modules/common/components/LoadingSpinner.styles.js
@@ -1,9 +1,0 @@
-// @flow
-
-import styled from 'styled-components'
-import Spinner from 'react-spinkit'
-
-export const CenteredSpinner = styled(Spinner)`
-  margin-top: 50px;
-  text-align: center;
-`
