+++ src/modules/layout/components/HeaderActionBar.js
@@ -3,8 +3,31 @@
 import React from 'react'
 import Link from 'redux-first-router-link'
 
-import { ActionItems } from './HeaderActionBar.styles'
 import HeaderActionItem from '../HeaderActionItem'
+import styled from 'styled-components'
+
+const ActionItems = styled.div`
+  justify-content: flex-end;
+  
+  & > *,
+  & img {
+    width: calc(0.8 * ${props => props.theme.dimensions.headerHeightLarge}px);
+    height: calc(0.8 * ${props => props.theme.dimensions.headerHeightLarge}px);
+    box-sizing: border-box;
+    cursor: pointer;
+    
+    @media ${props => props.theme.dimensions.smallViewport} {
+      width: calc(0.8 * ${props => props.theme.dimensions.headerHeightSmall}px);
+      height: calc(0.8 * ${props => props.theme.dimensions.headerHeightSmall}px);
+    }
+  }
+  
+  & img {
+    box-sizing: border-box;
+    padding: 22%;
+    object-fit: contain;
+  }
+`
 
 type PropsType = {
   className?: string,
