--- src/modules/layout/components/HeaderLanguageSelectorItem.styles.js
@@ -1,7 +1,0 @@
-// @flow
-
-import styled from 'styled-components'
-
-export const InactiveImage = styled.img`
-  color: ${props => props.theme.colors.textSecondaryColor};
-`
