+++ src/modules/layout/components/HeaderNavigationItem.js
@@ -3,8 +3,35 @@
 import React from 'react'
 import ReactTooltip from 'react-tooltip'
 
-import { InactiveNavigationItem, ActiveNavigationItem } from './HeaderNavigationItem.styles'
 import type { Action } from 'redux-first-router'
+import styled from 'styled-components'
+import Link from 'redux-first-router-link'
+
+const NavigationItem = styled(Link)`
+  ${props => props.theme.helpers.removeLinkHighlighting};
+  flex: 1;
+  color: ${props => props.theme.colors.textColor};
+  font-size: 1.1em;
+  font-weight: 400;
+  text-align: center;
+  text-transform: uppercase;
+
+  @media ${props => props.theme.dimensions.smallViewport} {
+    font-size: 0.9em;
+  }
+`
+
+const InactiveNavigationItem = styled(NavigationItem.withComponent('span'))`
+  color: ${props => props.theme.colors.textSecondaryColor};
+`
+
+const ActiveNavigationItem = styled(NavigationItem)`
+  ${props => props.selected
+  ? `font-weight: 700;`
+  : `:hover {
+        font-weight: 700;
+       }`}
+`
 
 type PropsType = {
   text: string,
